requirements.txt
python-telegram-bot>=20.0a0
python-dotenv>=0.20.0
